Move MigrateToNewDirStructure to common utils.
